articleObjectModel: rename body-html->html
Think this makes this a little clearer. The html we have in the model
includes html and body tags, and is ready for unstyled rendering in a
web page.

Later when we strip off the html and body tags and add it into the
template, we call that variable body-html
[endlessm/eos-sdk#2638]