engine: Don't use the Engine's content_path for loading search results
Extract it from the EKN ID instead. This removes a small dependency
for global state. At the same time, cache it in the engine so that
we don't hit the disk on every object lookup.

This is a slight behavior change, so fix up the tests and make sure they
all work correctly.