engine: Move the content_path property to be a passed-in domain
For the global search provider process, we want to be able to have
the ability to search multiple domains. Refactor this so that we can
pass in the domain to a search query, and find the database from that.

Since we have a lot of existing code that uses "get_objects_by_query"
without passing in a domain, and since that isn't easy to refactor,
have a "default_domain" property for now which specifies a fallback
domain in case no other exists. In the future, we should have a better
separation layer between the high-level application code and the
low-level engine code where the application code always passes in a
domain.