xapianQuery: use more descriptive names for query types
We have two query type, one of which before was the unnamed "general"
option and the other of which was called prefix search. I think it
will help to adopt some more descriptive names here--delimited
and incremental search. See
http://en.wikipedia.org/wiki/Delimited_search
http://en.wikipedia.org/wiki/Incremental_search
for a rough explanation, but I think it better encapsulates the
two query types we are offering
[endlessm/eos-sdk#2771]