engine: add type query parameter
Rather than have two distinct parameters for regular queries and
prefix search, refactored get_objects_by_query to take in a query
string and explicit "type" field for the type of search to preform.

These types are right now either QUERY_TYPE_DELIMETED (the old
general search) and QUERY_TYPE_INCREMENTAL.

This killed the ability to request a autocomplete and delimted
search at the same time, but don't think that made a lick of sense.
[endlessm/eos-sdk#2771]