engine: set cutoff value internally
A correct cutoff value for xapian depends more on query construction
than on anything else. Since we construct our queries internally in
the engine, we should set the cutoff value there as well instead of
making it public api of the engine.

This allows us to loosen the cutoff value for title only searches,
which gives better results
[endlessm/eos-sdk#2771]