js: move bulk of javascript from /overrides to /js
We are going to switch things up by having a javascript runtime
which calls into a C private library when necessary. This will get
us out of the business of fighting upstream gjs to try to make
javascript library, which it is sadly not built for.

So under the new scheme we will be able to drop all of our overrides
hacks, and have a simple single import path for all of our javascript
code. Should be easier to follow and easier to set up a testing
environment.

I'm moving everything that was in the overrides directory to a js/app
directory, to keep it clear that its a separate module from the search
library. We want our code in js/app to use whats in js/search, but never
the reverse.
[endlessm/eos-sdk#3108]