engine: use gio style callbacks for query api
Rather than the node style (err, results) functions, we will follow
the gio style of a asynchronous callbacks. To do this added a
AsyncTask class which basically imitates the non-introspectable
gtask, in a javascript friendly way. get_object_by_id and
get_objects_by_query will be split into async and finish functions,
as are the other async helper functions in the search library.
[endlessm/eos-sdk#2855]