Categories are sets in the database
This introduces compatibility code for 2.3-style app.jsons to change
their list of sections into objects in the database. It uses the database
runtime overlay functionality introduced in a previous commit, to create
ContentObjectModels in the database with the special tag "home page".
The presenter does a query for all objects with this tag at startup in
order to populate the cards on the home page.

Note that while cards with the "home page" are currently assumed to be
sets, there is no reason why documents couldn't also be on the home page
in the future.

For this compatibility layer we do assume, for performance reasons, that
runtime database objects do appear on the home page.

[endlessm/eos-sdk#3337]