Generate EKN ID if model doesn't have one
For sorting, filtering, and fast retrieval, we want to key models by
their EKN IDs. This is not a problem in production code, but for a lot
of tests we just call new ContentObjectModel() if we don't care, and
those models don't have EKN IDs.

To solve this problem, we generate an EKN ID with domain "none" if no ID
is given at construct time.

[endlessm/eos-sdk#4000]