Don't cancel ShardFile.init_async()
This operation will get cancelled by the shell every time the user types
a new letter in the desktop search box. If it's cancelled while in the
init_async() operation, then the ShardFile object will be in an error
state, and as per the GAsyncInitable docs, this spoils the object for any
future use: subsequent method calls are programmer errors.

Instead, we make the init_async() part of the operation uncancellable.

https://phabricator.endlessm.com/T10595