domain: Stop looking at the domain in the EKN ID
We now use the default domain everywhere -- the domain in an EKN ID is
now considered legacy and is ignored. The new, proper format of an EKN
ID is: ekn:///hash/resource, with the resource being optional. Old IDs
with domains will still work, however, the domain part will be ignored.