domain: Synchronously load domains
There really isn't anything we can do while an app is starting up, other
than show the splash / loading screen.

The only "big" thing this loses is the ability to load multi-shard; now,
we do it serially. This is because JavaScript considers parallelism and
asynchronous programming to be one and the same. We need to be able to
dispatch out multiple tasks, and have the JS call stack wait on the
results. This would require some C code to manage a thread pool.