Return results info object from engine queries
Since we'll be returning more metadata about queries, such as upper bound
on number of results, in a subsequent commit, this refactors
Engine.get_objects_by_query() to return a results info dict as its second
return value.

Previously it returned a "get more results query" in this position, which
is now available as the value of the dictionary's "more_results" key.

https://phabricator.endlessm.com/T11277