Never add more than the num_desired results to content group
When we have a filter present, we naively request three times
the original num_desired value, to account for the filter.
This results in us sometimes getting back more results than
we originally asked for. In that case we don't want to add those
extra results to the models map.

The implication of this is that we now need to move the creation
of the more_results_query out of engine.js and into Selection.Xapian
because only there do we have context for how many results were
really added to the arrangement and therefore from what offset
we need to begin our next query.

https://phabricator.endlessm.com/T13535