domain: load shards and link tables during init
After looking through the current state of the code for a while,
I don't think the public load_sync function being called all over
the place was really accomplishing anything.

All the public functions of domain would need to call load_sync
immediately to be usable, and some would forget and implicitly rely
on the fact that it had already been called.

Since engine will "lazy load" the domains only when needed, there's
no benefit to holding off on loading shards on domains until a method
has been called.

I think it makes sense to just load up domain shard content right
away in _init. When we port this to C and objects cannot error in
their initializers, we'll move to a GInitable.

https://phabricator.endlessm.com/T13418