domain: Store fetched models in a GPtrArray in the same order
We now care about the order in which Xapian returned values,
so we need to store them in the same order all the way throughout.

Using a GList with the asynchronous function meant that we could
receive the returned models out of order.

Also added a ScatterTaskCountdown object, which tracks "scattered"
callbacks, with a callback to be called once all the child callbacks
have completed (similar to Promise.all())