ekncontent: Don't ignore error conditions
This commit fixes a couple of calls to
eknc_domain_process_subscription() so they return on error rather than
reusing the already set GError pointer (which leads to an assertion
failure).

It also fixes the error handling of eknc_domain_get_impl() so that it
returns NULL if the initialization failed.