Support subscriptions from extensions
Starting with SDK2, we will be able to separate the content
from  the app bundle. The content  will be distributed  and
updated using flatpak extensions.

During this transition we will still need to support legacy
apps  that come bundled with  the content. For now  we keep
the current path as the mount point for the extensions:

   /app/share/ekn/data/$APP_ID/com.endlessm.subscriptions

The structure of this directory will also remain the same:

   com.endlessm.subscriptions:
    ├── uuid-1
    │   ├── manifest.json
    │   ├── sha256-1.shard
    │   └── sha256-2.shard
    ├── uuid-2
    │   ├── manifest.json
    │   └── sha256-3.shard
    └── uuid-3
        ├── manifest.json
        └── sha256-4.shard

But, we no longer depend on subscription.json file. we keep
the file though as a  transitioning step,  e.g. so we don't
mess with the code that removes the previous links from the
~/.local/share/com.endlessm.subscriptions directory.

https://phabricator.endlessm.com/T18489