ekncontent: drop subscriptions.json
Since we moved to flatpak and dropped the downloader
we no longer need to differenciate the subscriptions
from different apps. Apps own all the subscriptions
that are visible now.

Therefore, we no  longer need to parse and process
this subscriptions files anymore, but, we keep the
the  get_subscriptions API so we  can still handle
the removal of migrated apps to remove symlinks.

https://phabricator.endlessm.com/T19428