Remove old style Xapian query string
We don't use it anywhere, and it's here just for backward compatibility
reasons that ceased to matter once we moved to Flatpak runtimes and we
dropped Xapian-bridge.

Make the QueryObject a little less magic

The QueryObject caches a lot of the results, and the query strings have
side effects that are poorly documented or explained.

Let's make this a little bit less magic, and add API to extract every
query-related string explicitly.