Query both original and corrected terms
If we have a series of spelling corrections, we should merge them with
the original query, to get a query encompassing all the potential
results space.

This allows us to remove a lot of cruft from the QueryObject class, and
to streamline its API so that we can ask for the appropriate query
strings depending on the context.