ekncontent: Throw error if the content is empty
Since the content can be provided both from several
extensions or bundled with the app, we have certain
flexibility when we load the content, but too much.

Add an explicit error type for when no content was
found after all the subscriptions import steps are
done, and throw it when needed.

https://phabricator.endlessm.com/T20343