ekncontent: Don't put query parameters through hash tables
Previously we put all the parameters from an EkncQueryObject into a hash
table, then took them out again in order to create XapianQuery and
XapianEnquire objects in eknc_database_manager_query(). Now we can create
a XapianQuery and configure a XapianEnquire directly from the
EkncQueryObject, and get rid of the hash tables.

Some query parameters are always given the same value. In particular, for
a fix-query operation, all the query parameters that we previously put
into the hash table were the same. We encode them directly in the
fix-query operation in the database manager.

The language is a property of the EkncDomain, not the EkncQueryObject, so
we pass it in separately.

The query's limit and offset are also passed in separately to
xapian_enquire_get_mset().

https://phabricator.endlessm.com/T20373