ekncontent: Move XapianQuery parsing into EkncQueryObject
This adds a new private method, eknc_query_object_get_query(), that houses
the logic previously in EkncDatabaseManager to create a XapianQuery object
from an EkncQueryObject.

We make the methods that returned QueryParser strings into static functions.
The intention is to stop creating queries with string manipulation, so these
functions shouldn't be exposed in the API. One of them
(eknc_query_object_get_query_string()) wasn't used at all, so it is removed.

This also allows short-circuiting a bit more logic when the literal-query
property is set (which is only for debugging, though.)

It also allows writing a more comprehensive set of tests to catch
regressions in the queries we send to Xapian. This unearthed some oddities
which will have to be investigated in the future.

https://phabricator.endlessm.com/T20373