ekncontent: Strip newlines from Xapian's corrected terms
Some time ago, Olly fixed the database building code to not produce
spelling dictionaries with newlines on the end of each word:
https://github.com/endlessm/endless-content-factory/pull/755

However, not all databases have been rebuilt since then. We need to
manually remove newlines from Xapian's corrected terms if they are still
present. Otherwise, a spurious 'title:\n' will get added to the query,
and that will not be recognized as a prefix and therefore always search
for articles with 'title' in the title.

https://phabricator.endlessm.com/T20730